온갓 버그를 수정

+++ Assets/Main Menu/BGM/BGMManager.cs
@@ -51,7 +51,7 @@ namespace SDJK.MainMenu
 
         void Refresh()
         {
-            float lastTime = 0;
+            double lastTime = 0;
             if (bgm != null && !bgm.isRemoved)
             {
                 bgm.padeOut = true;
